Add files via upload
I made a quick level for us to use, ill expand on it but its just for simple use for now, i think its the sample scene that actually contains the entire scene, but I would recommend downloading all of the files since they contain all of the assets that I used + may leave in  so it will be good to have them. Just drag it into the bottom project folder in the unity engine where it says assets and it should load for everyone.